+++ nab-consul/src/main/java/ru/hh/nab/consul/ConsulMetricsTracker.java
@@ -2,7 +2,7 @@ package ru.hh.nab.consul;
 
 import ru.hh.consul.monitoring.ClientEventCallback;
 import ru.hh.metrics.Counters;
-import ru.hh.metrics.StatsDSender;
+import ru.hh.metrics.MetricsSender;
 import ru.hh.metrics.Tag;
 import static ru.hh.metrics.Tag.APP_TAG_NAME;
 
@@ -15,11 +15,11 @@ public class ConsulMetricsTracker implements ClientEventCallback {
   private final String serviceName;
   private final Counters requestCounters;
 
-  public ConsulMetricsTracker(String serviceName, StatsDSender statsDSender) {
+  public ConsulMetricsTracker(String serviceName, MetricsSender metricsSender) {
     this.serviceName = serviceName;
     this.requestCounters = new Counters(500);
 
-    statsDSender.sendPeriodically(() -> statsDSender.sendCounters(CONSUL_REQUESTS_METRIC, requestCounters));
+    metricsSender.sendPeriodically(() -> metricsSender.sendCounters(CONSUL_REQUESTS_METRIC, requestCounters));
   }
 
   @Override