+++ nab-kafka/src/main/java/ru/hh/nab/kafka/util/ConfigProvider.java
@@ -17,9 +17,9 @@ import org.apache.kafka.clients.consumer.ConsumerConfig;
 import static org.apache.kafka.clients.consumer.ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG;
 import static org.apache.kafka.clients.consumer.ConsumerConfig.MAX_POLL_INTERVAL_MS_CONFIG;
 import org.apache.kafka.clients.producer.ProducerConfig;
-import ru.hh.metrics.StatsDSender;
+import ru.hh.metrics.MetricsSender;
 import static ru.hh.nab.common.qualifier.NamedQualifier.SERVICE_NAME;
-import ru.hh.nab.kafka.monitoring.KafkaStatsDReporter;
+import ru.hh.nab.kafka.monitoring.KafkaMetricsReporter;
 import ru.hh.platform.utils.properties.PropertiesUtils;
 
 public class ConfigProvider {
@@ -54,21 +54,21 @@ public class ConfigProvider {
   static final Predicate<Object> NAB_SETTING_PREDICATE = key -> ((String) key).startsWith(NAB_SETTING + ".");
   private static final Set<String> SUPPORTED_PROPERTIES = Set.of(
       SERVICE_NAME,
-      KafkaStatsDReporter.STATSD_INSTANCE_PROPERTY,
-      KafkaStatsDReporter.METRICS_ALLOWED,
-      KafkaStatsDReporter.METRICS_SEND_ALL
+      KafkaMetricsReporter.METRICS_SENDER_INSTANCE_PROPERTY,
+      KafkaMetricsReporter.METRICS_ALLOWED,
+      KafkaMetricsReporter.METRICS_SEND_ALL
   );
   private final String serviceName;
   private final String kafkaClusterName;
   private final Properties properties;
-  private final StatsDSender statsDSender;
+  private final MetricsSender metricsSender;
 
-  public ConfigProvider(String serviceName, String kafkaClusterName, Properties properties, StatsDSender statsDSender) {
+  public ConfigProvider(String serviceName, String kafkaClusterName, Properties properties, MetricsSender metricsSender) {
     this.serviceName = serviceName;
     this.kafkaClusterName = kafkaClusterName;
     this.properties = new Properties();
     this.properties.putAll(properties);
-    this.statsDSender = statsDSender;
+    this.metricsSender = metricsSender;
   }
 
   private static void checkNames(Map<String, ?> nonNabSettings, Set<String> supportedNames, String type) {
@@ -224,26 +224,24 @@ public class ConfigProvider {
 
     String metricReporters = ofNullable(properties.get(CommonClientConfigs.METRIC_REPORTER_CLASSES_CONFIG))
         .map(Object::toString)
-        .orElseGet(KafkaStatsDReporter.class::getName);
+        .orElseGet(KafkaMetricsReporter.class::getName);
     properties.put(CommonClientConfigs.METRIC_REPORTER_CLASSES_CONFIG, metricReporters);
 
-    String metricsSendAll = ofNullable(properties.get(KafkaStatsDReporter.METRICS_SEND_ALL))
+    String metricsSendAll = ofNullable(properties.get(KafkaMetricsReporter.METRICS_SEND_ALL))
         .map(Object::toString)
         .orElseGet(Boolean.FALSE::toString);
-    properties.put(KafkaStatsDReporter.METRICS_SEND_ALL, metricsSendAll);
+    properties.put(KafkaMetricsReporter.METRICS_SEND_ALL, metricsSendAll);
 
-    String enabledMetrics = ofNullable(properties.get(KafkaStatsDReporter.METRICS_ALLOWED))
+    String enabledMetrics = ofNullable(properties.get(KafkaMetricsReporter.METRICS_ALLOWED))
         .map(Object::toString)
         .orElse("");
-    properties.put(KafkaStatsDReporter.METRICS_ALLOWED, enabledMetrics);
+    properties.put(KafkaMetricsReporter.METRICS_ALLOWED, enabledMetrics);
 
-    // TODO Remove when we leave Okmeter monitoring
-    // Okmeter doesn't provide precision better than once a minute
     properties.put(ConsumerConfig.METRICS_SAMPLE_WINDOW_MS_CONFIG, 60000);
     // Approximation, kafka defaults are 30000ms for sample window and 2 for num samples
     properties.put(ConsumerConfig.METRICS_NUM_SAMPLES_CONFIG, 4);
-    // A workaround to support single instance of StatsD client
-    properties.put(KafkaStatsDReporter.STATSD_INSTANCE_PROPERTY, this.statsDSender);
+    // A workaround to support single instance of MetricsSender client
+    properties.put(KafkaMetricsReporter.METRICS_SENDER_INSTANCE_PROPERTY, this.metricsSender);
     return properties;
   }